refactor: dao 를 interface로 추상화
