+++ src/main/java/roomescape/console/view/Command.java
@@ -1,0 +1,7 @@
+package roomescape.console.view;
+
+public enum Command {
+    POST,
+    GET,
+    DELETE;
+}
