+++ src/main/java/roomescape/console/view/Menu.java
@@ -2,11 +2,12 @@ package roomescape.console.view;
 
 public enum Menu {
     RESERVATION,
-    RESERVATION_TIME;
+    RESERVATION_TIME,
+    END;
 
     public static Menu from(int index) {
         for (Menu value : Menu.values()) {
-            if (value.ordinal() == index) {
+            if (value.ordinal() == (index - 1)) {
                 return value;
             }
         }
