+++ src/main/java/roomescape/dao/ReservationTimeDao.java
@@ -1,46 +1,13 @@
 package roomescape.dao;
 
-import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.jdbc.core.RowMapper;
-import org.springframework.jdbc.core.namedparam.BeanPropertySqlParameterSource;
-import org.springframework.jdbc.core.namedparam.SqlParameterSource;
-import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
-import org.springframework.stereotype.Component;
 import roomescape.domain.ReservationTime;
 
 import java.util.List;
 
-@Component
-public class ReservationTimeDao {
+public interface ReservationTimeDao {
+    ReservationTime create(ReservationTime reservationTime);
 
-    private final JdbcTemplate jdbcTemplate;
-    private final SimpleJdbcInsert jdbcInsert;
+    List<ReservationTime> findAll();
 
-    public ReservationTimeDao(JdbcTemplate jdbcTemplate) {
-        this.jdbcTemplate = jdbcTemplate;
-        this.jdbcInsert = new SimpleJdbcInsert(jdbcTemplate)
-                .withTableName("RESERVATION_TIME")
-                .usingGeneratedKeyColumns("id");
-    }
-
-    private final RowMapper<ReservationTime> reservationTimeRowMapper = (resultSet, rowNum) -> new ReservationTime(
-            resultSet.getInt("id"),
-            resultSet.getTime("start_at").toLocalTime()
-    );
-
-    public ReservationTime create(ReservationTime reservationTime) {
-        SqlParameterSource params = new BeanPropertySqlParameterSource(reservationTime);
-        long id = jdbcInsert.executeAndReturnKey(params).longValue();
-        return new ReservationTime(id, reservationTime.getStartAt());
-    }
-
-    public List<ReservationTime> findAll() {
-        String sql = "select * from reservation_time";
-        return jdbcTemplate.query(sql, reservationTimeRowMapper);
-    }
-
-    public void delete(long id) {
-        String sql = "delete from reservation_time where id = ?";
-        jdbcTemplate.update(sql, id);
-    }
+    void delete(long id);
 }
