+++ src/main/java/roomescape/dao/ReservationTimeDaoImpl.java
@@ -1,0 +1,49 @@
+package roomescape.dao;
+
+import org.springframework.jdbc.core.JdbcTemplate;
+import org.springframework.jdbc.core.RowMapper;
+import org.springframework.jdbc.core.namedparam.BeanPropertySqlParameterSource;
+import org.springframework.jdbc.core.namedparam.SqlParameterSource;
+import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
+import org.springframework.stereotype.Component;
+import roomescape.domain.ReservationTime;
+
+import java.util.List;
+
+@Component
+public class ReservationTimeDaoImpl implements ReservationTimeDao {
+
+    private final JdbcTemplate jdbcTemplate;
+    private final SimpleJdbcInsert jdbcInsert;
+
+    public ReservationTimeDaoImpl(JdbcTemplate jdbcTemplate) {
+        this.jdbcTemplate = jdbcTemplate;
+        this.jdbcInsert = new SimpleJdbcInsert(jdbcTemplate)
+                .withTableName("RESERVATION_TIME")
+                .usingGeneratedKeyColumns("id");
+    }
+
+    private final RowMapper<ReservationTime> reservationTimeRowMapper = (resultSet, rowNum) -> new ReservationTime(
+            resultSet.getInt("id"),
+            resultSet.getTime("start_at").toLocalTime()
+    );
+
+    @Override
+    public ReservationTime create(ReservationTime reservationTime) {
+        SqlParameterSource params = new BeanPropertySqlParameterSource(reservationTime);
+        long id = jdbcInsert.executeAndReturnKey(params).longValue();
+        return new ReservationTime(id, reservationTime.getStartAt());
+    }
+
+    @Override
+    public List<ReservationTime> findAll() {
+        String sql = "select * from reservation_time";
+        return jdbcTemplate.query(sql, reservationTimeRowMapper);
+    }
+
+    @Override
+    public void delete(long id) {
+        String sql = "delete from reservation_time where id = ?";
+        jdbcTemplate.update(sql, id);
+    }
+}
